+++ codas-kit/src/main.rs
@@ -1,0 +1,7 @@
+use clap::Parser;
+use codas_kit::args::Args;
+
+fn main() {
+    let args = Args::parse();
+    args.execute();
+}